Fixed a bunch of issues with the new beta build.
Changed the "title_menu" in the strings back to "Menu" because it is the menu title in the menu.

Fixed an issue where setting a selection in the fitness location activity would stop the app if there wasn't a location.

Added a check to the setCheckbox tick in EquipmentActivity because the call to teh db was getting back results before the menuCheckBox was being initialized.

Added "(DEBUG)" to the about screen if we are using the debug build.

Stopped the app from crashing if we open the settings in an older android version.